WIP, make index (homepage) show more dev value props and features
