+++ website/src/css/pages/_customers.scss
@@ -1,0 +1,12 @@
+.customers-page {
+    &__header-and-hero {
+        background: url('/customers-page-bg.svg') no-repeat;
+        background-size: cover;
+    }
+    .list-group-item {
+        background-color: rgba(255, 255, 255, 0.75);
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+    }
+}
