+++ website/src/css/styles.scss
@@ -250,6 +250,7 @@ table {
 @import 'pages/browser';
 @import 'pages/blog';
 @import 'pages/blog-post';
+@import 'pages/customers';
 @import 'pages/_case_studies';
 @import 'pages/_white_papers';
 @import 'pages/content';
